+++ src/led.c
@@ -1,42 +1,9 @@
-/************************************************************************************//**
-* \file         Demo/ARMCM33_STM32H5_Nucleo_H563ZI_GCC/Prog/led.c
-* \brief        LED driver source file.
-* \ingroup      Prog_ARMCM33_STM32H5_Nucleo_H563ZI_GCC
-* \internal
-*----------------------------------------------------------------------------------------
-*                          C O P Y R I G H T
-*----------------------------------------------------------------------------------------
-*   Copyright (c) 2024  by Feaser    http://www.feaser.com    All rights reserved
-*
-*----------------------------------------------------------------------------------------
-*                            L I C E N S E
-*----------------------------------------------------------------------------------------
-* This file is part of OpenBLT. OpenBLT is free software: you can redistribute it and/or
-* modify it under the terms of the GNU General Public License as published by the Free
-* Software Foundation, either version 3 of the License, or (at your option) any later
-* version.
-*
-* OpenBLT is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
-* without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
-* PURPOSE. See the GNU General Public License for more details.
-*
-* You have received a copy of the GNU General Public License along with OpenBLT. It 
-* should be located in ".\Doc\license.html". If not, contact Feaser to obtain a copy.
-*
-* \endinternal
-****************************************************************************************/
-
 /****************************************************************************************
 * Include files
 ****************************************************************************************/
-#include "header.h"                                    /* generic header               */
+#include "header.h"                                   
 
-
-/****************************************************************************************
-* Macro definitions
-****************************************************************************************/
-/** \brief Toggle interval time in milliseconds. */
-#define LED_TOGGLE_MS  (500)
+#define LED_TOGGLE_MS  (500)          //led toggle interval in milliseconds
 
 
 /************************************************************************************//**
@@ -46,13 +13,10 @@
 ****************************************************************************************/
 void LedInit(void)
 {
-  /* Note that the initialization of the LED GPIO pin is done in HAL_MspInit(). All that
-   * is left to do here is to make sure the LED is turned off after initialization.
-   */
   HAL_GPIO_WritePin(GPIOB, GPIO_PIN_0, GPIO_PIN_RESET);
   HAL_GPIO_WritePin(GPIOF, GPIO_PIN_4, GPIO_PIN_RESET);
   HAL_GPIO_WritePin(GPIOG, GPIO_PIN_4, GPIO_PIN_RESET);
-} /*** end of LedInit ***/
+}
 
 
 /************************************************************************************//**
